requirements.txt
Django==4.2.4
psycopg==3.1.10
psycopg2-binary==2.8.5
pytest==7.4.0
pytest-django==4.5.2
requests==2.31.0
tzdata==2023.3
Pillow==10.0.1
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
unidecode==1.3.7
python-decouple==3.8
passlib==1.7.4
djangorestframework-simplejwt==5.3.0